requirements.txt
flask
flask-login
mysql-connector-python
gunicorn
werkzeug